Display clearout offer for large old lists
[MAILPOET-2919]
