+++ assets/js/src/subscribers/importExport/import/step_input_validation.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { withRouter } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
+import OfferClearout from 'subscribers/importExport/import/offer_clearout';
 import InitialQuestion from './step_input_validation/initial_question.jsx';
 import WrongSourceBlock from './step_input_validation/wrong_source_block.jsx';
-import OldListBlock from './step_input_validation/old_list_block.jsx';
 import LastSentQuestion from './step_input_validation/last_sent_question.jsx';
 
 function StepInputValidation({ stepMethodSelectionData, history }) {
@@ -50,7 +50,7 @@ function StepInputValidation({ stepMethodSelectionData, history }) {
       )}
 
       {importSource === 'existing-list' && lastSent === 'notRecently' && (
-        <OldListBlock />
+        <OfferClearout />
       )}
     </div>
   );
