+++ assets/js/src/subscribers/importExport/import/step_offer_clearout.tsx
@@ -1,32 +1,7 @@
 import React from 'react';
-import Button from 'common/button/button';
-import MailPoet from 'mailpoet';
+import { RouteComponentProps } from 'react-router-dom';
+import OfferClearout from './offer_clearout';
 
-type Props = {
-  history: History,
-}
-
-export default ({ history }: Props) => (
-  <div className="mailpoet-offer-clearout-step__container">
-    <p>{MailPoet.I18n.t('offerClearoutText1')}</p>
-    <p>{MailPoet.I18n.t('offerClearoutText2')}</p>
-    <p>{MailPoet.I18n.t('offerClearoutText3')}</p>
-    <p>
-      <Button
-        dimension="small"
-        variant="dark"
-        target="_blank"
-        href="https://clearout.io/?ref=mailpoet"
-      >
-        {MailPoet.I18n.t('tryClearout')}
-      </Button>
-      <Button
-        onClick={() => (history.push('step_method_selection'))}
-        variant="link-dark"
-        dimension="small"
-      >
-        {MailPoet.I18n.t('clearoutGotIt')}
-      </Button>
-    </p>
-  </div>
+export default ({ history }: RouteComponentProps) => (
+  <OfferClearout onProceed={() => history.push('step_method_selection')} />
 );
